获取来访者预约等候中的订单

+++ src/main/java/cn/edu/tongji/healper/indto/ConsultTimeInDto.java
@@ -4,6 +4,8 @@ import lombok.Data;
 
 @Data
 public class ConsultTimeInDto {
+
     Integer orderId;
-    Integer time;
+
+    Long time;
 }
