+++ src/main/java/cn/edu/tongji/healper/service/ConsultService.java
@@ -2,7 +2,7 @@ package cn.edu.tongji.healper.service;
 
 
 public interface ConsultService {
-    Boolean startConsultation(Integer orderId, Integer startTime);
+    Boolean startConsultation(Integer orderId, Long startTime);
 
-    Boolean endConsultation(Integer orderId, Integer endTime);
+    Boolean endConsultation(Integer orderId, Long endTime);
 }
